requirements.txt
boto3==1.37.33
botocore==1.37.33
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
jmespath==1.0.1
lxml==5.3.2
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
requests==2.32.3
s3transfer==0.11.4
six==1.17.0
urllib3==2.4.0
